requirements.txt
colorama==0.3.7
dj-database-url==0.4.1
Django==1.10
gunicorn==19.6.0
Markdown==2.6.6
psycopg2==2.6.1
py==1.4.31
pytest==2.9.2
pytest-django==2.9.1
pytest-env==0.6.0
pytest-mock==1.1
unittest2py3k==0.5.1
whitenoise==3.2
django-configurations==1.0.0
django-dotenv==1.4.1
django-extensions==1.7.2
